LayoutTests/http/tests/security/cross-origin-css-4-expected.txt

CONSOLE MESSAGE: Did not parse stylesheet at 'http://127.0.0.1:8000/security/resources/xorigincss3.html' because its MIME type was invalid.

PASS Testing cross-origin and MIME behavior for CSS. 
PASS xorigincss3.html should not be loaded, even though it is same-origin 
